Organize module hierarchy
