+++ middleware/index.js
@@ -1,1 +1,2 @@
+module.exports.asyncTry = require('./asyncTry');
 module.exports.requestParser = require('./requestParser');
